Behaviour working for all but down to the last few nodes. .G factor (which prevents the same node from being chosen over and over) still to be implemented.

+++ clusterModel.h
@@ -22,7 +22,7 @@ typedef struct ClusterNodes{
 typedef struct ClusterModel{
 	NetArch netA;
 	NodeArch nodeA;
-	int numCluster;
+	float numCluster;
 	float p;
 	ClusterNodes clusterN;
 } ClusterModel;
